mongowrite更新
要素更新追加するの忘れてました

+++ ソースコード/test/tensorflow00/predictionService.py
@@ -78,13 +78,13 @@ P_TWENAVE = makePredictionModel(c.TWENAVE)
 P_TIME = GetDate()
 
 # 予測値だけを取り出す処理
-P_OPEN = P_OPEN[:,0].astype(np.float64)
-P_CLOSE = P_CLOSE[:,0].astype(np.float64)
-P_HIGH = P_HIGH[:,0].astype(np.float64)
-P_LOW = P_LOW[:,0].astype(np.float64)
-P_FIVEAVE = P_FIVEAVE[:0].astype(np.float64)
-P_TENAVE = P_TENAVE[:0].astype(np.float64)
-P_TWENAVE = P_TWENAVE[:0].astype(np.float64)
+P_OPEN = P_OPEN[:,0].astype(np.float32)
+P_CLOSE = P_CLOSE[:,0].astype(np.float32)
+P_HIGH = P_HIGH[:,0].astype(np.float32)
+P_LOW = P_LOW[:,0].astype(np.float32)
+P_FIVEAVE = P_FIVEAVE[:0].astype(np.float32)
+P_TENAVE = P_TENAVE[:0].astype(np.float32)
+P_TWENAVE = P_TWENAVE[:0].astype(np.float32)
 
 if (check.iloc[last,0] != time):
     # 辞書キーの作成
